sample: coap_client: Use network manager and move to net
The coap_client sample now uses the network
manager instead of directly interacting with the
nRF91 modem. Thus the sample now also supports
the nRF70 DK. Additionally printk statements have
been converted to use the log system.

The sample is also moved to the ned folder.
