--- samples/cellular/coap_client/sample.yaml
@@ -1,10 +1,0 @@
-sample:
-  name: nRF CoAP client sample
-tests:
-  sample.cellular.coap_client:
-    build_only: true
-    integration_platforms:
-      - nrf9160dk_nrf9160_ns
-      - nrf9161dk_nrf9161_ns
-    platform_allow: nrf9160dk_nrf9160_ns nrf9161dk_nrf9161_ns
-    tags: ci_build
