--- samples/cellular/coap_client/src/main.c
@@ -1,302 +1,0 @@
-/*
- * Copyright (c) 2019 Nordic Semiconductor ASA
- *
- * SPDX-License-Identifier: LicenseRef-Nordic-5-Clause
- */
-
-#include <stdio.h>
-#include <string.h>
-
-#include <zephyr/kernel.h>
-#include <zephyr/net/coap.h>
-#include <zephyr/net/socket.h>
-#include <modem/lte_lc.h>
-#include <modem/nrf_modem_lib.h>
-#include <zephyr/random/rand32.h>
-
-#define APP_COAP_SEND_INTERVAL_MS 5000
-#define APP_COAP_MAX_MSG_LEN 1280
-#define APP_COAP_VERSION 1
-
-static int sock;
-static struct pollfd fds;
-static struct sockaddr_storage server;
-static uint16_t next_token;
-
-static uint8_t coap_buf[APP_COAP_MAX_MSG_LEN];
-
-/**@brief Resolves the configured hostname. */
-static int server_resolve(void)
-{
-	int err;
-	struct addrinfo *result;
-	struct addrinfo hints = {
-		.ai_family = AF_INET,
-		.ai_socktype = SOCK_DGRAM
-	};
-	char ipv4_addr[NET_IPV4_ADDR_LEN];
-
-	err = getaddrinfo(CONFIG_COAP_SERVER_HOSTNAME, NULL, &hints, &result);
-	if (err != 0) {
-		printk("ERROR: getaddrinfo failed %d\n", err);
-		return -EIO;
-	}
-
-	if (result == NULL) {
-		printk("ERROR: Address not found\n");
-		return -ENOENT;
-	}
-
-	/* IPv4 Address. */
-	struct sockaddr_in *server4 = ((struct sockaddr_in *)&server);
-
-	server4->sin_addr.s_addr =
-		((struct sockaddr_in *)result->ai_addr)->sin_addr.s_addr;
-	server4->sin_family = AF_INET;
-	server4->sin_port = htons(CONFIG_COAP_SERVER_PORT);
-
-	inet_ntop(AF_INET, &server4->sin_addr.s_addr, ipv4_addr,
-		  sizeof(ipv4_addr));
-	printk("IPv4 Address found %s\n", ipv4_addr);
-
-	/* Free the address. */
-	freeaddrinfo(result);
-
-	return 0;
-}
-
-/**@brief Initialize the CoAP client */
-static int client_init(void)
-{
-	int err;
-
-	sock = socket(AF_INET, SOCK_DGRAM, IPPROTO_UDP);
-	if (sock < 0) {
-		printk("Failed to create CoAP socket: %d.\n", errno);
-		return -errno;
-	}
-
-	err = connect(sock, (struct sockaddr *)&server,
-		      sizeof(struct sockaddr_in));
-	if (err < 0) {
-		printk("Connect failed : %d\n", errno);
-		return -errno;
-	}
-
-	/* Initialize FDS, for poll. */
-	fds.fd = sock;
-	fds.events = POLLIN;
-
-	/* Randomize token. */
-	next_token = sys_rand32_get();
-
-	return 0;
-}
-
-/**@brief Handles responses from the remote CoAP server. */
-static int client_handle_get_response(uint8_t *buf, int received)
-{
-	int err;
-	struct coap_packet reply;
-	const uint8_t *payload;
-	uint16_t payload_len;
-	uint8_t token[8];
-	uint16_t token_len;
-	uint8_t temp_buf[16];
-
-	err = coap_packet_parse(&reply, buf, received, NULL, 0);
-	if (err < 0) {
-		printk("Malformed response received: %d\n", err);
-		return err;
-	}
-
-	payload = coap_packet_get_payload(&reply, &payload_len);
-	token_len = coap_header_get_token(&reply, token);
-
-	if ((token_len != sizeof(next_token)) ||
-	    (memcmp(&next_token, token, sizeof(next_token)) != 0)) {
-		printk("Invalid token received: 0x%02x%02x\n",
-		       token[1], token[0]);
-		return 0;
-	}
-
-	if (payload_len > 0) {
-		snprintf(temp_buf, MIN(payload_len, sizeof(temp_buf)), "%s", payload);
-	} else {
-		strcpy(temp_buf, "EMPTY");
-	}
-
-	printk("CoAP response: code: 0x%x, token 0x%02x%02x, payload: %s\n",
-	       coap_header_get_code(&reply), token[1], token[0], temp_buf);
-
-	return 0;
-}
-
-/**@biref Send CoAP GET request. */
-static int client_get_send(void)
-{
-	int err;
-	struct coap_packet request;
-
-	next_token++;
-
-	err = coap_packet_init(&request, coap_buf, sizeof(coap_buf),
-			       APP_COAP_VERSION, COAP_TYPE_NON_CON,
-			       sizeof(next_token), (uint8_t *)&next_token,
-			       COAP_METHOD_GET, coap_next_id());
-	if (err < 0) {
-		printk("Failed to create CoAP request, %d\n", err);
-		return err;
-	}
-
-	err = coap_packet_append_option(&request, COAP_OPTION_URI_PATH,
-					(uint8_t *)CONFIG_COAP_RESOURCE,
-					strlen(CONFIG_COAP_RESOURCE));
-	if (err < 0) {
-		printk("Failed to encode CoAP option, %d\n", err);
-		return err;
-	}
-
-	err = send(sock, request.data, request.offset, 0);
-	if (err < 0) {
-		printk("Failed to send CoAP request, %d\n", errno);
-		return -errno;
-	}
-
-	printk("CoAP request sent: token 0x%04x\n", next_token);
-
-	return 0;
-}
-
-/**@brief Configures modem to provide LTE link. Blocks until link is
- * successfully established.
- */
-static void modem_configure(void)
-{
-#if defined(CONFIG_LTE_LINK_CONTROL)
-
-	int err;
-
-	printk("LTE Link Connecting ...\n");
-	err = lte_lc_init_and_connect();
-	__ASSERT(err == 0, "LTE link could not be established.");
-	printk("LTE Link Connected!\n");
-
-#endif /* defined(CONFIG_LTE_LINK_CONTROL) */
-}
-
-/* Returns 0 if data is available.
- * Returns -EAGAIN if timeout occured and there is no data.
- * Returns other, negative error code in case of poll error.
- */
-static int wait(int timeout)
-{
-	int ret = poll(&fds, 1, timeout);
-
-	if (ret < 0) {
-		printk("poll error: %d\n", errno);
-		return -errno;
-	}
-
-	if (ret == 0) {
-		/* Timeout. */
-		return -EAGAIN;
-	}
-
-	if ((fds.revents & POLLERR) == POLLERR) {
-		printk("wait: POLLERR\n");
-		return -EIO;
-	}
-
-	if ((fds.revents & POLLNVAL) == POLLNVAL) {
-		printk("wait: POLLNVAL\n");
-		return -EBADF;
-	}
-
-	if ((fds.revents & POLLIN) != POLLIN) {
-		return -EAGAIN;
-	}
-
-	return 0;
-}
-
-int main(void)
-{
-	int64_t next_msg_time = APP_COAP_SEND_INTERVAL_MS;
-	int err, received;
-
-	printk("The nRF CoAP client sample started\n");
-
-	err = nrf_modem_lib_init();
-	if (err) {
-		printk("Modem library initialization failed, error: %d\n", err);
-		return 0;
-	}
-
-	modem_configure();
-
-	if (server_resolve() != 0) {
-		printk("Failed to resolve server name\n");
-		return 0;
-	}
-
-	if (client_init() != 0) {
-		printk("Failed to initialize CoAP client\n");
-		return 0;
-	}
-
-	next_msg_time = k_uptime_get();
-
-	while (1) {
-		if (k_uptime_get() >= next_msg_time) {
-			if (client_get_send() != 0) {
-				printk("Failed to send GET request, exit...\n");
-				break;
-			}
-
-			next_msg_time += APP_COAP_SEND_INTERVAL_MS;
-		}
-
-		int64_t remaining = next_msg_time - k_uptime_get();
-
-		if (remaining < 0) {
-			remaining = 0;
-		}
-
-		err = wait(remaining);
-		if (err < 0) {
-			if (err == -EAGAIN) {
-				continue;
-			}
-
-			printk("Poll error, exit...\n");
-			break;
-		}
-
-		received = recv(sock, coap_buf, sizeof(coap_buf), MSG_DONTWAIT);
-		if (received < 0) {
-			if (errno == EAGAIN || errno == EWOULDBLOCK) {
-				printk("socket EAGAIN\n");
-				continue;
-			} else {
-				printk("Socket error, exit...\n");
-				break;
-			}
-		}
-
-		if (received == 0) {
-			printk("Empty datagram\n");
-			continue;
-		}
-
-		err = client_handle_get_response(coap_buf, received);
-		if (err < 0) {
-			printk("Invalid response, exit...\n");
-			break;
-		}
-	}
-
-	(void)close(sock);
-
-	return 0;
-}
