+++ samples/net/coap_client/CMakeLists.txt
