+++ samples/net/coap_client/boards/nrf7002dk_nrf5340_cpuapp.conf
@@ -1,0 +1,43 @@
+#
+# Copyright (c) 2023 Nordic Semiconductor ASA
+#
+# SPDX-License-Identifier: LicenseRef-Nordic-5-Clause
+#
+
+# Configuration file for nRF7002 DK
+# This file is merged with prj.conf in the application folder, and options
+# set here will take precedence if they are present in both files.
+
+# General
+CONFIG_POSIX_CLOCK=y
+CONFIG_SYSTEM_WORKQUEUE_STACK_SIZE=4096
+CONFIG_LOG_BUFFER_SIZE=4096
+CONFIG_HW_STACK_PROTECTION=y
+
+# Wi-Fi
+CONFIG_WIFI=y
+CONFIG_WIFI_NRF700X=y
+CONFIG_WIFI_MGMT_EXT=y
+CONFIG_WIFI_CREDENTIALS=y
+CONFIG_WIFI_CREDENTIALS_BACKEND_SETTINGS=y
+CONFIG_WIFI_CREDENTIALS_STATIC=y
+
+# WPA
+CONFIG_WPA_SUPP=y
+
+# NET sockets
+CONFIG_NET_L2_ETHERNET=y
+CONFIG_NET_TCP_WORKQ_STACK_SIZE=2048
+CONFIG_NET_UDP=y
+CONFIG_NET_NATIVE=y
+CONFIG_NET_DHCPV4=y
+CONFIG_NET_MGMT_EVENT_STACK_SIZE=4096
+CONFIG_NET_RX_STACK_SIZE=4096
+
+# DNS
+CONFIG_DNS_RESOLVER=y
+
+# Zephyr NET Connection Manager Connectivity layer.
+CONFIG_L2_WIFI_CONNECTIVITY=y
+CONFIG_L2_WIFI_CONNECTIVITY_AUTO_DOWN=n
+CONFIG_L2_WIFI_CONNECTIVITY_AUTO_CONNECT=n
