+++ samples/net/coap_client/boards/nrf9161dk_nrf9161_ns.conf
@@ -1,0 +1,28 @@
+#
+# Copyright (c) 2023 Nordic Semiconductor ASA
+#
+# SPDX-License-Identifier: LicenseRef-Nordic-5-Clause
+#
+
+# Configuration file for nRF9160 DK
+# This file is merged with prj.conf in the application folder, and options
+# set here will take precedence if they are present in both files.
+
+# General
+CONFIG_HW_STACK_PROTECTION=y
+CONFIG_HW_ID_LIBRARY_SOURCE_IMEI=y
+
+# Modem related configurations
+CONFIG_NRF_MODEM_LIB_NET_IF_AUTO_START=n
+CONFIG_NRF_MODEM_LIB_ON_FAULT_APPLICATION_SPECIFIC=y
+
+# Disable Duplicate Address Detection (DAD)
+# due to not being properly implemented for offloaded interfaces.
+CONFIG_NET_IPV6_NBR_CACHE=n
+CONFIG_NET_IPV6_MLD=n
+
+# Zephyr NET Connection Manager and Connectivity layer.
+CONFIG_NET_CONNECTION_MANAGER_MONITOR_STACK_SIZE=1024
+CONFIG_LTE_CONNECTIVITY=y
+CONFIG_LTE_CONNECTIVITY_AUTO_DOWN=n
+CONFIG_LTE_CONNECTIVITY_AUTO_CONNECT=n
