+++ samples/net/coap_client/boards/qemu_x86.conf
