+++ samples/net/coap_client/prj.conf
@@ -4,25 +4,15 @@
 # SPDX-License-Identifier: LicenseRef-Nordic-5-Clause
 #
 # General config
-CONFIG_HEAP_MEM_POOL_SIZE=4096
+CONFIG_LOG=y
+CONFIG_LOG_BACKEND_UART=y
 
 # Networking
 CONFIG_NETWORKING=y
-CONFIG_NET_NATIVE=n
-CONFIG_NET_SOCKETS_OFFLOAD=y
 CONFIG_NET_SOCKETS=y
 CONFIG_NET_SOCKETS_POSIX_NAMES=y
-
-# LTE link control
-CONFIG_LTE_LINK_CONTROL=y
-
-# Modem library
-CONFIG_NRF_MODEM_LIB=y
-CONFIG_NRF_MODEM_LIB_TRACE=n
-
-# AT Host
-CONFIG_UART_INTERRUPT_DRIVEN=y
-CONFIG_AT_HOST_LIBRARY=y
+CONFIG_NET_IPV4=y
+CONFIG_NET_CONNECTION_MANAGER=y
 
 # CoAP
 CONFIG_COAP=y
