+++ samples/net/coap_client/sample.yaml
@@ -1,0 +1,20 @@
+sample:
+  name: nRF CoAP client sample
+tests:
+  sample.cellular.coap_client:
+    build_only: true
+    build_on_all: true
+    integration_platforms:
+      - nrf9160dk_nrf9160_ns
+      - nrf9161dk_nrf9161_ns
+      - thingy91_nrf9160_ns
+      - nrf7002dk_nrf5340_cpuapp
+    platform_allow:
+      - nrf9160dk_nrf9160_ns
+      - nrf9161dk_nrf9161_ns
+      - thingy91_nrf9160_ns
+      - nrf7002dk_nrf5340_cpuapp
+    extra_configs:
+      - CONFIG_WIFI_CREDENTIALS_STATIC_SSID="test-ssid"
+      - CONFIG_WIFI_CREDENTIALS_STATIC_PASSWORD="test-password"
+    tags: ci_build
