+++ samples/net/coap_client/src/main.c
@@ -1,0 +1,396 @@
+/*
+ * Copyright (c) 2019 Nordic Semiconductor ASA
+ *
+ * SPDX-License-Identifier: LicenseRef-Nordic-5-Clause
+ */
+
+#include <stdio.h>
+#include <string.h>
+
+#include <zephyr/kernel.h>
+#include <zephyr/net/coap.h>
+#include <zephyr/net/socket.h>
+#include <zephyr/net/conn_mgr_connectivity.h>
+#include <zephyr/net/conn_mgr_monitor.h>
+#include <zephyr/random/rand32.h>
+#include <zephyr/logging/log.h>
+
+LOG_MODULE_REGISTER(coap_client_sample, CONFIG_COAP_CLIENT_SAMPLE_LOG_LEVEL);
+
+/* Macros used to subscribe to specific Zephyr NET management events. */
+#define L4_EVENT_MASK (NET_EVENT_L4_CONNECTED | NET_EVENT_L4_DISCONNECTED)
+#define CONN_LAYER_EVENT_MASK (NET_EVENT_CONN_IF_FATAL_ERROR)
+
+#define APP_COAP_SEND_INTERVAL_MS 5000
+#define APP_COAP_MAX_MSG_LEN 1280
+#define APP_COAP_VERSION 1
+
+/* Zephyr NET management event callback structures. */
+static struct net_mgmt_event_callback l4_cb;
+static struct net_mgmt_event_callback conn_cb;
+
+static int sock;
+static struct pollfd fds;
+static struct sockaddr_storage server;
+static uint16_t next_token;
+
+static uint8_t coap_buf[APP_COAP_MAX_MSG_LEN];
+
+K_MUTEX_DEFINE(network_connected_lock);
+K_CONDVAR_DEFINE(network_connected);
+static bool is_connected;
+
+/**
+ * @brief Resolves the configured hostname.
+ *
+ * @return 0 on success, error code otherwise.
+ */
+static int server_resolve(void)
+{
+	int err;
+	struct addrinfo *result;
+	struct addrinfo hints = {
+		.ai_family = AF_INET,
+		.ai_socktype = SOCK_DGRAM
+	};
+	char ipv4_addr[NET_IPV4_ADDR_LEN];
+
+	err = getaddrinfo(CONFIG_COAP_SERVER_HOSTNAME, NULL, &hints, &result);
+	if (err != 0) {
+		LOG_ERR("ERROR: getaddrinfo failed %d", err);
+		return -EIO;
+	}
+
+	if (result == NULL) {
+		LOG_ERR("ERROR: Address not found");
+		return -ENOENT;
+	}
+
+	/* IPv4 Address. */
+	struct sockaddr_in *server4 = ((struct sockaddr_in *)&server);
+
+	server4->sin_addr.s_addr =
+		((struct sockaddr_in *)result->ai_addr)->sin_addr.s_addr;
+	server4->sin_family = AF_INET;
+	server4->sin_port = htons(CONFIG_COAP_SERVER_PORT);
+
+	inet_ntop(AF_INET, &server4->sin_addr.s_addr, ipv4_addr,
+		  sizeof(ipv4_addr));
+	LOG_INF("IPv4 Address found %s", ipv4_addr);
+
+	/* Free the address. */
+	freeaddrinfo(result);
+
+	return 0;
+}
+
+/**
+ * @brief Initialize the CoAP client
+ *
+ * @return 0 on success, error code otherwise.
+ */
+static int client_init(void)
+{
+	int err;
+
+	sock = socket(AF_INET, SOCK_DGRAM, IPPROTO_UDP);
+	if (sock < 0) {
+		LOG_ERR("Failed to create CoAP socket: %d.", errno);
+		return -errno;
+	}
+
+	err = connect(sock, (struct sockaddr *)&server,
+		      sizeof(struct sockaddr_in));
+	if (err < 0) {
+		LOG_ERR("Connect failed : %d", errno);
+		return -errno;
+	}
+
+	/* Initialize FDS, for poll. */
+	fds.fd = sock;
+	fds.events = POLLIN;
+
+	/* Randomize token. */
+	next_token = sys_rand32_get();
+
+	return 0;
+}
+
+/**
+ * @brief Handles responses from the remote CoAP server.
+ *
+ * @return 0 on success, error code otherwise.
+ */
+static int client_handle_get_response(uint8_t *buf, int received)
+{
+	int err;
+	struct coap_packet reply;
+	const uint8_t *payload;
+	uint16_t payload_len;
+	uint8_t token[8];
+	uint16_t token_len;
+	uint8_t temp_buf[16];
+
+	err = coap_packet_parse(&reply, buf, received, NULL, 0);
+	if (err < 0) {
+		LOG_ERR("Malformed response received: %d", err);
+		return err;
+	}
+
+	payload = coap_packet_get_payload(&reply, &payload_len);
+	token_len = coap_header_get_token(&reply, token);
+
+	if ((token_len != sizeof(next_token)) ||
+	    (memcmp(&next_token, token, sizeof(next_token)) != 0)) {
+		LOG_ERR("Invalid token received: 0x%02x%02x",
+		       token[1], token[0]);
+		return 0;
+	}
+
+	if (payload_len > 0) {
+		snprintf(temp_buf, MIN(payload_len, sizeof(temp_buf)), "%s", payload);
+	} else {
+		strcpy(temp_buf, "EMPTY");
+	}
+
+	LOG_INF("CoAP response: code: 0x%x, token 0x%02x%02x, payload: %s",
+	       coap_header_get_code(&reply), token[1], token[0], temp_buf);
+
+	return 0;
+}
+
+/**
+ * @brief Send CoAP GET request.
+ *
+ * @return 0 on success, error code otherwise.
+ */
+static int client_get_send(void)
+{
+	int err;
+	struct coap_packet request;
+
+	next_token++;
+
+	err = coap_packet_init(&request, coap_buf, sizeof(coap_buf),
+			       APP_COAP_VERSION, COAP_TYPE_NON_CON,
+			       sizeof(next_token), (uint8_t *)&next_token,
+			       COAP_METHOD_GET, coap_next_id());
+	if (err < 0) {
+		LOG_ERR("Failed to create CoAP request, %d", err);
+		return err;
+	}
+
+	err = coap_packet_append_option(&request, COAP_OPTION_URI_PATH,
+					(uint8_t *)CONFIG_COAP_RESOURCE,
+					strlen(CONFIG_COAP_RESOURCE));
+	if (err < 0) {
+		LOG_ERR("Failed to encode CoAP option, %d", err);
+		return err;
+	}
+
+	err = send(sock, request.data, request.offset, 0);
+	if (err < 0) {
+		LOG_ERR("Failed to send CoAP request, %d", errno);
+		return -errno;
+	}
+
+	LOG_INF("CoAP request sent: token 0x%04x", next_token);
+
+	return 0;
+}
+
+/**
+ * @brief Waits until data is available on the socket
+ *
+ * @return 0 on success
+ * @return -EAGAIN if timeout occurred and there is no data.
+ * @return negative error code in case of poll error.
+ */
+static int wait(int timeout)
+{
+	int ret = poll(&fds, 1, timeout);
+
+	if (ret < 0) {
+		LOG_ERR("poll error: %d", errno);
+		return -errno;
+	}
+
+	if (ret == 0) {
+		/* Timeout. */
+		return -EAGAIN;
+	}
+
+	if ((fds.revents & POLLERR) == POLLERR) {
+		LOG_ERR("wait: POLLERR");
+		return -EIO;
+	}
+
+	if ((fds.revents & POLLNVAL) == POLLNVAL) {
+		LOG_ERR("wait: POLLNVAL");
+		return -EBADF;
+	}
+
+	if ((fds.revents & POLLIN) != POLLIN) {
+		return -EAGAIN;
+	}
+
+	return 0;
+}
+
+/**
+ * @brief Waits for the establishment of a network connection.
+ */
+static void wait_for_network(void)
+{
+	k_mutex_lock(&network_connected_lock, K_FOREVER);
+	if (!is_connected) {
+		LOG_INF("Waiting for network connectivity");
+		k_condvar_wait(&network_connected, &network_connected_lock, K_FOREVER);
+	}
+	k_mutex_unlock(&network_connected_lock);
+}
+
+/**
+ * @brief Infinite CoAP request loop
+ */
+static void coap_loop(void)
+{
+	int64_t next_msg_time = APP_COAP_SEND_INTERVAL_MS;
+	int err, received;
+
+	wait_for_network();
+
+	if (server_resolve() != 0) {
+		LOG_ERR("Failed to resolve server name");
+		return;
+	}
+
+	if (client_init() != 0) {
+		LOG_ERR("Failed to initialize CoAP client");
+		return;
+	}
+
+	next_msg_time = k_uptime_get();
+
+	while (1) {
+		wait_for_network();
+
+		if (k_uptime_get() >= next_msg_time) {
+			if (client_get_send() != 0) {
+				LOG_ERR("Failed to send GET request, exit...");
+				break;
+			}
+
+			next_msg_time = k_uptime_get() + APP_COAP_SEND_INTERVAL_MS;
+		}
+
+		int64_t remaining = next_msg_time - k_uptime_get();
+
+		if (remaining < 0) {
+			remaining = 0;
+		}
+
+		err = wait(remaining);
+		if (err < 0) {
+			if (err == -EAGAIN) {
+				continue;
+			}
+
+			LOG_ERR("Poll error...");
+			continue;
+		}
+
+		received = recv(sock, coap_buf, sizeof(coap_buf), MSG_DONTWAIT);
+		if (received < 0) {
+			LOG_ERR("Socket error, error: %d", errno);
+		}
+
+		if (received == 0) {
+			LOG_ERR("Empty datagram");
+			continue;
+		}
+
+		err = client_handle_get_response(coap_buf, received);
+		if (err < 0) {
+			LOG_ERR("Invalid response...");
+			continue;
+		}
+	}
+
+	(void)close(sock);
+}
+
+static void l4_event_handler(struct net_mgmt_event_callback *cb,
+			     uint32_t event,
+			     struct net_if *iface)
+{
+	switch (event) {
+	case NET_EVENT_L4_CONNECTED:
+		LOG_INF("Network connectivity established");
+		k_mutex_lock(&network_connected_lock, K_FOREVER);
+		is_connected = true;
+		k_condvar_signal(&network_connected);
+		k_mutex_unlock(&network_connected_lock);
+		break;
+	case NET_EVENT_L4_DISCONNECTED:
+		LOG_INF("Network connectivity lost");
+		k_mutex_lock(&network_connected_lock, K_FOREVER);
+		is_connected = false;
+		k_mutex_unlock(&network_connected_lock);
+		break;
+	default:
+		/* Don't care */
+		return;
+	}
+}
+
+static void connectivity_event_handler(struct net_mgmt_event_callback *cb,
+				       uint32_t event,
+				       struct net_if *iface)
+{
+	if (event == NET_EVENT_CONN_IF_FATAL_ERROR) {
+		LOG_ERR("NET_EVENT_CONN_IF_FATAL_ERROR");
+		k_fatal_halt(K_ERR_KERNEL_OOPS);
+		return;
+	}
+}
+
+int main(void)
+{
+	int err;
+
+	LOG_INF("The nRF CoAP client sample started");
+
+	/* Setup handler for Zephyr NET Connection Manager events. */
+	net_mgmt_init_event_callback(&l4_cb, l4_event_handler, L4_EVENT_MASK);
+	net_mgmt_add_event_callback(&l4_cb);
+
+	/* Setup handler for Zephyr NET Connection Manager Connectivity layer. */
+	net_mgmt_init_event_callback(&conn_cb, connectivity_event_handler, CONN_LAYER_EVENT_MASK);
+	net_mgmt_add_event_callback(&conn_cb);
+
+	/* Bring all network interfaces up.
+	 * Wi-Fi or LTE depending on the board that the sample was built for.
+	 */
+	LOG_INF("Bringing network interface up and connecting to the network");
+	err = conn_mgr_all_if_up(true);
+	if (err) {
+		LOG_ERR("conn_mgr_all_if_up, error: %d", err);
+		return err;
+	}
+
+	err = conn_mgr_all_if_connect(true);
+	if (err) {
+		printk("conn_mgr_all_if_connect, error: %d\n", err);
+		return err;
+	}
+
+
+	/* Start the infinite coap request loop. */
+	coap_loop();
+
+	/* If we get here an error occurred. Halt the system. */
+	k_fatal_halt(K_ERR_KERNEL_OOPS);
+
+	return 0;
+}
